Split config to elastic and main
