+++ src/main/java/wiki/config/JsonFileConfiguration.java
@@ -1,0 +1,4 @@
+package wiki.config;
+
+public class JsonFileConfiguration {
+}
