+++ src/main/java/wiki/data/relations/LinkAndParenthesisRelationExtractor.java
@@ -3,7 +3,7 @@ package wiki.data.relations;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import wiki.data.obj.LinkParenthesisPair;
-import wiki.utils.LangConfiguration;
+import wiki.config.LangConfiguration;
 import wiki.utils.parsers.WikiPageParser;
 
 import java.util.ArrayList;
