+++ src/main/java/wiki/persistency/JsonFileAPI.java
@@ -1,0 +1,4 @@
+package wiki.persistency;
+
+public class JsonFileAPI {
+}
