+++ src/test/java/wiki/frequency/TestWordFrequencyAndRep.java
@@ -4,8 +4,7 @@ import com.google.gson.Gson;
 import org.junit.jupiter.api.Test;
 import wiki.TestUtils;
 import wiki.WikiToElasticMain;
-import wiki.utils.LangConfiguration;
-import wiki.utils.WikiToElasticConfiguration;
+import wiki.config.MainConfiguration;
 
 import java.io.FileNotFoundException;
 import java.io.FileReader;
@@ -17,7 +16,7 @@ public class TestWordFrequencyAndRep {
     @Test
     public void testCountLineFrequency() throws FileNotFoundException {
         String testConfig = Objects.requireNonNull(WikiToElasticMain.class.getClassLoader().getResource("test_config.json")).getFile();
-        WikiToElasticConfiguration config = GSON.fromJson(new FileReader(testConfig), WikiToElasticConfiguration.class);
+        MainConfiguration config = GSON.fromJson(new FileReader(testConfig), MainConfiguration.class);
 
         WordFrequencyAndRepresentation.initResources(config.getLang());
         WordFrequencyAndRepresentation wfar = new WordFrequencyAndRepresentation();
